Move instance avatar
