+++ lib/instance/widgets/instance_list_entry.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:thunder/instance/utils/navigate_instance.dart';
-import 'package:thunder/instance/widgets/instance_avatar.dart';
+import 'package:thunder/shared/avatars/instance_avatar.dart';
 import 'package:thunder/utils/instance.dart';
 import 'package:thunder/utils/numbers.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
